Text: fixed last issues with empty text box if no text
Text: fixed last issues with empty text box if no text and right text size on tool change

+++ client/src/app/services/manager/shortcut-manager.service.ts
@@ -162,11 +162,13 @@ export class ShortcutManagerService {
     }
 
     onDeleteKeyDown(event: KeyboardEvent): void {
-        event.preventDefault();
         if (!this.isShortcutAllowed()) {
             return;
         }
-        this.clipboardService.deleteSelection();
+        if (this.toolManager.currentTool instanceof RectangleSelectionService || this.toolManager.currentTool instanceof EllipseSelectionService) {
+            event.preventDefault();
+            this.clipboardService.deleteSelection();
+        }
     }
 
     onMinusKeyDown(): void {
